// apps/web/src/api/requests/services.gen.ts
export class DefaultService {
  /**
   * @returns unknown The Health Check is successful
   * @throws ApiError
   */
  public static healthControllerCheck(): CancelablePromise<HealthControllerCheckResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/health",
      errors: {
        503: "The Health Check is not successful",
      },
    });
  }
}

export class AuthService {
  /**
   * @param data The data for the request.
   * @param data.requestBody
   * @returns LoginOutputDto Success
   * @throws ApiError
   */
  public static login(data: LoginData): CancelablePromise<LoginResponse> {
    return __request(OpenAPI, {
      method: "POST",
      url: "/api/auth/login",
      body: data.requestBody,
      mediaType: "application/json",
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @returns unknown
   * @throws ApiError
   */
  public static logout(): CancelablePromise<LogoutResponse> {
    return __request(OpenAPI, {
      method: "POST",
      url: "/api/auth/logout",
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.requestBody
   * @returns GenericMessageDto Success
   * @throws ApiError
   */
  public static initiatePasswordReset(
    data: InitiatePasswordResetData,
  ): CancelablePromise<InitiatePasswordResetResponse> {
    return __request(OpenAPI, {
      method: "POST",
      url: "/api/auth/password-reset/initiate",
      body: data.requestBody,
      mediaType: "application/json",
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.requestBody
   * @returns GenericMessageDto Success
   * @throws ApiError
   */
  public static completePasswordReset(
    data: CompletePasswordResetData,
  ): CancelablePromise<CompletePasswordResetResponse> {
    return __request(OpenAPI, {
      method: "POST",
      url: "/api/auth/password-reset/complete",
      body: data.requestBody,
      mediaType: "application/json",
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.requestBody
   * @returns GenericMessageDto Success
   * @throws ApiError
   */
  public static completeOnboarding(
    data: CompleteOnboardingData,
  ): CancelablePromise<CompleteOnboardingResponse> {
    return __request(OpenAPI, {
      method: "POST",
      url: "/api/auth/onboarding/complete",
      body: data.requestBody,
      mediaType: "application/json",
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.requestBody
   * @returns GenericMessageDto Success
   * @throws ApiError
   */
  public static linkOauthAccount(
    data: LinkOauthAccountData,
  ): CancelablePromise<LinkOauthAccountResponse> {
    return __request(OpenAPI, {
      method: "POST",
      url: "/api/auth/link-account",
      body: data.requestBody,
      mediaType: "application/json",
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @returns unknown
   * @throws ApiError
   */
  public static google(): CancelablePromise<GoogleResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/auth/providers/google",
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.state
   * @returns unknown
   * @throws ApiError
   */
  public static authControllerGoogleCallback(
    data: AuthControllerGoogleCallbackData,
  ): CancelablePromise<AuthControllerGoogleCallbackResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/auth/providers/google/callback",
      query: {
        state: data.state,
      },
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @returns unknown
   * @throws ApiError
   */
  public static microsoft(): CancelablePromise<MicrosoftResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/auth/providers/microsoft",
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.state
   * @returns unknown
   * @throws ApiError
   */
  public static authControllerMicrosoftCallback(
    data: AuthControllerMicrosoftCallbackData,
  ): CancelablePromise<AuthControllerMicrosoftCallbackResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/auth/providers/microsoft/callback",
      query: {
        state: data.state,
      },
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }
}
// ...
export class MapsService {
  /**
   * @param data The data for the request.
   * @param data.page
   * @param data.recordsPerPage
   * @returns MapOutputPageDto Success
   * @throws ApiError
   */
  public static findManyMaps(
    data: FindManyMapsData,
  ): CancelablePromise<FindManyMapsResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/maps",
      query: {
        page: data.page,
        recordsPerPage: data.recordsPerPage,
      },
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.requestBody
   * @returns MapOutputDto Success
   * @throws ApiError
   */
  public static createMap(
    data: CreateMapData,
  ): CancelablePromise<CreateMapResponse> {
    return __request(OpenAPI, {
      method: "POST",
      url: "/api/maps",
      body: data.requestBody,
      mediaType: "application/json",
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @returns MapOutputDto Success
   * @throws ApiError
   */
  public static findDefaultMap(): CancelablePromise<FindDefaultMapResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/maps/default",
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.id
   * @returns MapOutputDto Success
   * @throws ApiError
   */
  public static findOneMap(
    data: FindOneMapData,
  ): CancelablePromise<FindOneMapResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/maps/{id}",
      path: {
        id: data.id,
      },
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        404: "Not found",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.id
   * @param data.requestBody
   * @returns MapOutputDto Success
   * @throws ApiError
   */
  public static updateMap(
    data: UpdateMapData,
  ): CancelablePromise<UpdateMapResponse> {
    return __request(OpenAPI, {
      method: "PUT",
      url: "/api/maps/{id}",
      path: {
        id: data.id,
      },
      body: data.requestBody,
      mediaType: "application/json",
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        404: "Not found",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.id
   * @returns MapOutputDto Success
   * @throws ApiError
   */
  public static deleteMap(
    data: DeleteMapData,
  ): CancelablePromise<DeleteMapResponse> {
    return __request(OpenAPI, {
      method: "DELETE",
      url: "/api/maps/{id}",
      path: {
        id: data.id,
      },
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        404: "Not found",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @returns string Success
   * @throws ApiError
   */
  public static exportMaps(): CancelablePromise<ExportMapsResponse> {
    return __request(OpenAPI, {
      method: "POST",
      url: "/api/maps/export",
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }
}

export class AlarmsService {
  /**
   * @param data The data for the request.
   * @param data.page
   * @param data.recordsPerPage
   * @returns AlarmOutputPageDto Success
   * @throws ApiError
   */
  public static findManyAlarms(
    data: FindManyAlarmsData,
  ): CancelablePromise<FindManyAlarmsResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/alarms",
      query: {
        page: data.page,
        recordsPerPage: data.recordsPerPage,
      },
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.requestBody
   * @returns AlarmOutputDto Success
   * @throws ApiError
   */
  public static createAlarm(
    data: CreateAlarmData,
  ): CancelablePromise<CreateAlarmResponse> {
    return __request(OpenAPI, {
      method: "POST",
      url: "/api/alarms",
      body: data.requestBody,
      mediaType: "application/json",
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.id
   * @returns AlarmOutputDto Success
   * @throws ApiError
   */
  public static findOneAlarm(
    data: FindOneAlarmData,
  ): CancelablePromise<FindOneAlarmResponse> {
    return __request(OpenAPI, {
      method: "GET",
      url: "/api/alarms/{id}",
      path: {
        id: data.id,
      },
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        404: "Not found",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.id
   * @param data.requestBody
   * @returns AlarmOutputDto Success
   * @throws ApiError
   */
  public static updateAlarm(
    data: UpdateAlarmData,
  ): CancelablePromise<UpdateAlarmResponse> {
    return __request(OpenAPI, {
      method: "PUT",
      url: "/api/alarms/{id}",
      path: {
        id: data.id,
      },
      body: data.requestBody,
      mediaType: "application/json",
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        404: "Not found",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @param data The data for the request.
   * @param data.id
   * @returns AlarmOutputDto Success
   * @throws ApiError
   */
  public static deleteAlarm(
    data: DeleteAlarmData,
  ): CancelablePromise<DeleteAlarmResponse> {
    return __request(OpenAPI, {
      method: "DELETE",
      url: "/api/alarms/{id}",
      path: {
        id: data.id,
      },
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        404: "Not found",
        500: "Internal Server Error",
      },
    });
  }

  /**
   * @returns string Success
   * @throws ApiError
   */
  public static exportAlarms(): CancelablePromise<ExportAlarmsResponse> {
    return __request(OpenAPI, {
      method: "POST",
      url: "/api/alarms/export",
      errors: {
        401: "Unauthorized",
        403: "Forbidden",
        500: "Internal Server Error",
      },
    });
  }
}